Blocking methods now rethrow errors sent to on_error.

+++ projects/scripts/travis-doxygen.sh
@@ -2,10 +2,12 @@
 
 set -e
 
-cd projects/doxygen/html
-git init
-git config user.name "Kirk Shoop"
-git config user.email "kirk.shoop@microsoft.com"
-git add *
-git commit -m "doxygen generated site"
-git push --force "https://${GH_TOKEN}@github.com/Reactive-Extensions/RxCpp.git" master:gh-pages
+if [ -n "$GH_TOKEN" ]; then
+    cd projects/doxygen/html
+    git init
+    git config user.name "Kirk Shoop"
+    git config user.email "kirk.shoop@microsoft.com"
+    git add *
+    git commit -m "doxygen generated site"
+    git push --force "https://${GH_TOKEN}@github.com/Reactive-Extensions/RxCpp.git" master:gh-pages
+fi
